remove GetSrcItem/GetDstItem variants.
Especially source, there may be multiple valid entries for a hash in the
same source. This method obscures this.

Dst is technically ok, but let's not let users not think about it.

In both cases, the user should decide if SrcID/DstID is the right
comparator. In some cases SrcItem/DstItem EqualKey() may be more
appropriate.